switched from int to long for shape

+++ src/main/java/org/ilastik/bdvsource/StructuredInfo.java
@@ -8,7 +8,7 @@ public class StructuredInfo {
     public static class ImageInfo {
         public String name;
         public String axes;
-        public int[] shape;
+        public long[] shape;
         public int lane_number;
         public String dtype;
         public String dataset_name;
